Add support for Java 6
Removed the use of NIO2 for file globbing, thereby allowing Java 6 to
work with boot. Used Spring's AntPathMatcher instead, which adds Spring
as a dependency. That may be more of a dependency than you want. Spring
is under the Apache License, so you should be able to copy just the
class you need, if that's a better choice.